--- src/app/signin-group/join-group/JoinGroupClient.module.scss
@@ -1,109 +1,0 @@
-.searchContainer {
-  position: relative;
-  width: 100%;
-
-  .inputWrapper {
-    position: relative;
-    display: flex;
-    align-items: center;
-    width: 100%;
-
-    .searchInput {
-      @include text-style(16, $black, light);
-
-      width: 100%;
-      padding: 1.4rem 0 1.3rem 1.3rem;
-      border: 0.05rem solid $gray;
-      border-radius: 0.6rem;
-      transition: border-color 0.3s ease;
-      background-color: $white;
-      z-index: 1;
-
-      &::placeholder {
-        @include text-style(18, $black);
-        font-weight: 300;
-      }
-
-      &:focus {
-        border-color: $primary;
-        outline: none;
-      }
-    }
-
-    .searchButton {
-      position: absolute;
-      z-index: 2;
-
-      width: 4.7rem;
-      right: 0;
-      border-radius: 0 0.5rem 0.5rem 0;
-      background-color: $gray;
-      border: none;
-      height: 100%;
-      cursor: pointer;
-      transition: $base-transition;
-
-      &:hover {
-        background-color: $primary;
-      }
-    }
-
-    &:focus-within {
-      .searchButton {
-        background-color: $primary;
-      }
-    }
-  }
-
-  .suggestions {
-    @include column-flexbox(center, start);
-
-    position: absolute;
-    top: 3.7rem;
-    width: 100%;
-    border: 0.05rem solid $primary;
-    background-color: $white;
-    border-radius: 0.6rem;
-    box-shadow: 0 0.2rem 0.8rem rgba(0, 0, 0, 0.1);
-    padding-top: 0.8rem;
-
-    .suggestionItem {
-      @include flexbox(start, center, 0.6rem);
-      @include text-style(16, $black, light);
-
-      width: 100%;
-      padding: 1rem;
-      cursor: pointer;
-      transition: $base-transition;
-      border-radius: 0.6rem;
-
-      &:hover {
-        background-color: $gray30;
-        color: $primary;
-      }
-    }
-
-    .highlight {
-      font-weight: 800;
-    }
-  }
-}
-
-@keyframes slideIn {
-  0% {
-    opacity: 0;
-    transform: translateY(-1rem);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.errorMessage {
-  @include text-style(16, red);
-  margin-top: 1rem;
-  margin-left: 1rem;
-
-  animation: slideIn 0.3s ease-out;
-}
