+++ src/components/[id]/CommentSection.module.scss
@@ -1,0 +1,11 @@
+.add-comment-form {
+  width: 100%;
+
+  .add-comment {
+    @include text-style(12, $black);
+    width: 100%;
+    &::placeholder {
+      @include text-style(12, $gray);
+    }
+  }
+}
