+++ src/components/[id]/LikeButton.tsx
