+++ src/components/myfamily/storage/SwipeableListItem.module.scss
@@ -1,0 +1,97 @@
+.draggableContent {
+  @include column-flexbox(start, start, 1.3rem);
+
+  width: 100%;
+  cursor: grab;
+  background-color: $white;
+  border-radius: 0.6rem;
+  position: relative;
+  user-select: none;
+  transition: transform 0.3s ease;
+
+  &-noticeButton {
+    position: absolute;
+    width: 0.3rem;
+    height: auto;
+    background-color: $gray20;
+    top: 0.5rem;
+    bottom: 0.5rem;
+    right: 1rem;
+    border-radius: 1rem;
+  }
+  &-button {
+    width: 100%;
+    margin: 1.8rem 0 1.7rem 2.4rem;
+    @include column-flexbox(start, start, 1.3rem);
+
+    &-title {
+      @include text-style(16, $black, bolder);
+    }
+
+    &-duration {
+      @include text-style(16, $black, light);
+    }
+  }
+}
+
+.draggableContent.showButton {
+  transform: translateX(-12.7rem);
+}
+
+.Buttons {
+  @include flexbox(center, center);
+  width: 11.7rem;
+  height: 8.2rem;
+  background-color: $gray20;
+  border: none;
+  border-radius: 0.6rem;
+  padding: 1.2rem 0 1rem;
+  position: absolute;
+  right: -12.7rem;
+  top: 50%;
+  transform: translateY(-50%);
+  transition: opacity 1s ease;
+  opacity: 0;
+  cursor: default;
+
+  & > *:not(:last-child) {
+    border-right: 0.1rem solid $gray;
+    border-right-color: $gray;
+  }
+}
+
+.Buttons.showButton {
+  opacity: 1;
+  cursor: pointer;
+}
+
+.button-cover {
+  @include flexbox(center, center);
+  height: 100%;
+  width: 100%;
+  border: 0.1rem solid transparent;
+
+  .button {
+    transition: $base-transition;
+  }
+
+  &:hover {
+    .revise {
+      filter: brightness(40%);
+    }
+    .delete {
+      color: red;
+    }
+  }
+}
+
+.addButton {
+  transition: $base-transition;
+  position: absolute;
+  right: 3.8rem;
+  bottom: 20.1rem;
+
+  &:hover {
+    filter: brightness(0.9);
+  }
+}
